FasterRCNN/requirements.txt
conda==4.9.0
Cython==0.29.21
matplotlib==3.3.2
numpy==1.18.5
opencv-python-headless==4.4.0.46
pandas==1.1.4
Pillow==7.2.0
pycocotools==2.0.2
torch==1.6.0
torchvision==0.7.0
tqdm==4.46.0